feat: 사용자 목록 조회 및 날짜 포멧 추가

+++ src/main/java/com/example/demo/common/dto/response/ResponseListBody.java
@@ -22,7 +22,7 @@ import java.util.List;
 public class ResponseListBody<T> implements ResponseBody {
 
     @Builder.Default
-    private Long totalCount = 0L;
+    private int totalCount = 0;
 
     @Builder.Default
     private int totalPage = 0;
